fix: TripEnded 이벤트 totalDistance 필드 타입 BigDecimal -> Long 수정

+++ support-service/src/main/java/com/gooddaytaxi/support/adapter/in/kafka/dto/TripEndedEventPayload.java
@@ -18,7 +18,7 @@ public record TripEndedEventPayload(
         LocalDateTime startTime,
         LocalDateTime endTime,
         Long totalDuration,
-        BigDecimal totalDistance,
+        Long totalDistance,
         Long finalFare
 ) {
 }
